pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "azul-plugin-mandiant-capa"
description = "Use FireEye's CAPA tool to detect capabilities in binaries."
dynamic = ["version"]
readme = "README.md"
license = "MIT"
license-files = ["license.md"]
authors = [
  { name="Azul"},
]
requires-python = ">=3.12"
classifiers=[
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "azul-runner>=9.0.38",
    "capstone>=5.0.6",
    "flare-capa>=9.3.0",
    "pefile>=2024.8.26",
]

[project.scripts]
azul-plugin-mandiant-capa = "azul_plugin_mandiant_capa.main:main"

[project.urls]
Documentation = "https://australiancybersecuritycentre.github.io/azul/"
Repository = "https://github.com/AustralianCyberSecurityCentre/azul-plugin-mandiant-capa"

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "only-version"

[tool.hatch.version.fallback]
version = "0.0.0"

[tool.hatch.build.targets.wheel]
packages = ["azul_plugin_mandiant_capa", "capa-rules", "capa-rules-cache", "capa-sigs"]

[dependency-groups]
dev = [
    "azul-runner[test-utils]>=9.0.38",
    "black>=26.1.0",
    "cart>=1.2.3",
    "flake8>=7.3.0",
    "flake8-bugbear>=25.11.29",
    "flake8-docstrings>=1.7.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-resource-usage>=1.0.0",
    "tox>=4.34.1",
]

[tool.ruff]
# formatter config
line-length = 119
indent-width = 4
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".env",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Flake8 linting equivalent but for ruff
[tool.ruff.lint]
select = [
  # isort
  "I",
  # Code styles
  "E4",
  "E7",
  "E9",
  # pyflakes
  "F",
  # flake8 bug bear
  "B",
  # flake8 doc strings
  "D",
  # bandit
  "S",
]
ignore = ["B008","C901","D104","D107","E203"]
exclude = [
    "tests/*",
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


[tool.ruff.lint.pydocstyle]
# Use the Google docstring convention
convention = "google"

# Make formatter match black
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
